refactor: move rules computation to SvgRootNode fix: apply opacity when fill-opacity or stroke-opacity is not present fix: gradient not considering viewport translation when gradientUnits == "objectBoundingBox
