+++ svg-to-compose/src/commonMain/kotlin/dev/tonholo/s2c/domain/svg/SvgEllipseNode.kt
@@ -45,10 +45,10 @@ fun SvgEllipseNode.asNode(
     return ImageVectorNode.Path(
         params = ImageVectorNode.Path.Params(
             fill = fillBrush(nodes),
-            fillAlpha = fillOpacity,
+            fillAlpha = fillOpacity ?: opacity,
             pathFillType = fillRule,
             stroke = strokeBrush(nodes),
-            strokeAlpha = strokeOpacity,
+            strokeAlpha = strokeOpacity ?: opacity,
             strokeLineCap = strokeLineCap,
             strokeLineJoin = strokeLineJoin,
             strokeMiterLimit = strokeMiterLimit,
